feat: added authentication jwt

+++ src/routes/memories.ts
@@ -3,9 +3,15 @@ import { prisma } from "../lib/prisma";
 import {z} from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-//listagem de memórias
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+    //listagem de memórias
+    app.get('/memories', async (request) => {
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy: {
                 createdAt: 'asc'
             }
@@ -18,8 +24,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
             }
         })
     })
-// listagem de uma memória específica
-    app.get('/memories/:id', async (request) => {
+    // listagem de uma memória específica
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         }) 
@@ -29,10 +35,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             },
         })
+
+        if(!memory.isPublic && memory.userId !== request.user.sub) {
+            reply.status(403).send({
+                error: 'Forbidden'
+            })
+        }
+
         return memory
     })
 
-// criação de memória 
+    // criação de memória 
     app.post('/memories', async (request) => {
         const bodySchema = z.object({
             content:z.string(),
@@ -47,14 +60,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 isPublic,
                 coverUrl,
-                userId: '0fd2b444-0e72-4dd8-bb04-901c7370f22c',
+                userId: request.user.sub,
             }
         })
         return memory
     })
 
-// atualização de memória 
-    app.put('/memories/:id', async (request) => {
+    // atualização de memória 
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         }) 
@@ -67,7 +80,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
         }) 
 
         const {content, isPublic, coverUrl} = bodySchema.parse(request.body)
-        const memoryUpdated = await prisma.memory.update({
+
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            },
+        })
+
+        if(memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
+
+       memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -77,15 +101,25 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 coverUrl,
             },
         })
-        return memoryUpdated
+        return memory
     })
 
-// remoção de memória 
-    app.delete('/memories/:id', async (request) => {
+    // remoção de memória 
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         }) 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            },
+        })
+
+        if(memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
          await prisma.memory.delete({
             where: {
                 id,
